+++ cluster/controller/reconciler_leader_test.go
@@ -1,0 +1,11 @@
+package controller
+
+import (
+	"testing"
+
+	"k8s.io/client-go/rest"
+)
+
+func testReconcilerLeader(t *testing.T, config *rest.Config) {
+
+}
